backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
optuna==3.4.0
shap==0.44.0
lime==0.2.0.1
imbalanced-learn==0.11.0

# Data Processing
scipy==1.11.4
statsmodels==0.14.0
pyarrow==14.0.1
h5py==3.10.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# HTTP & Async
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Caching & Queue
redis==5.0.1
celery==5.3.6

# File handling
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.7
Pillow==10.1.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.1
faker==20.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
cryptography==41.0.7
itsdangerous==2.1.2
jinja2==3.1.2
pyyaml==6.0.1
toml==0.10.2
click==8.1.7
rich==13.7.0
tqdm==4.66.1
python-dotenv==1.0.0
orjson==3.9.10
